requirements.txt
selenium>=4.0.0
webdriver-manager>=3.8.0
langchain==0.3.25
openai==1.82.0
langchain-openai==0.3.18
langgraph==0.4.7
python-dotenv==1.1.0
langchain-core==0.3.62
google-generativeai
langchain-google-genai
coverage
pytest-cov